% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logo.R
\name{create_logo}
\alias{create_logo}
\title{Customize useR! logo}
\usage{
create_logo(colour = "#ffc0cb", year = 2020, folder)
}
\arguments{
\item{colour}{hex code of the colour for use and "!"}

\item{year}{year of the conference}

\item{folder}{folder where to save the svg and PNG}
}
\description{
Customize useR! logo
}
\examples{
create_logo(colour = "#ffc0cb",
            year = 2020,
            folder = tempdir())
}
